+++ clikt-mordant/src/commonMain/kotlin/com/github/ajalt/clikt/testing/CliktTesting.kt
@@ -36,6 +36,9 @@ data class CliktCommandTestResult(
  * @param ansiLevel Defaults to no colored output; set to [AnsiLevel.TRUECOLOR] to include ANSI codes in the output.
  * @param width The width of the terminal, used to wrap text
  * @param height The height of the terminal
+ * @param hyperlinks Whether to enable hyperlink support in the terminal
+ * @param outputInteractive Whether the output is interactive
+ * @param inputInteractive Whether the input is interactive
  */
 fun CliktCommand.test(
     argv: String,
@@ -45,16 +48,22 @@ fun CliktCommand.test(
     ansiLevel: AnsiLevel = AnsiLevel.NONE,
     width: Int = 79,
     height: Int = 24,
+    hyperlinks: Boolean = ansiLevel != AnsiLevel.NONE,
+    outputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
+    inputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
 ): CliktCommandTestResult {
     val argvArray = CommandLineParser.tokenize(argv)
-    return test(argvArray, stdin, envvars, includeSystemEnvvars, ansiLevel, width, height)
+    return test(
+        argvArray, stdin, envvars, includeSystemEnvvars, ansiLevel, width, height,
+        hyperlinks, outputInteractive, inputInteractive
+    )
 }
 
 /**
  * Test this command, returning a result that captures the output and result status code.
  *
- * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed with [print] or
- * [println] is not.
+ * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed
+ * with [print] or [println] is not.
  *
  * @param argv The command line to send to the command
  * @param stdin Content of stdin that will be read by prompt options. Multiple inputs should be separated by `\n`.
@@ -64,6 +73,9 @@ fun CliktCommand.test(
  * @param ansiLevel Defaults to no colored output; set to [AnsiLevel.TRUECOLOR] to include ANSI codes in the output.
  * @param width The width of the terminal, used to wrap text
  * @param height The height of the terminal
+ * @param hyperlinks Whether to enable hyperlink support in the terminal
+ * @param outputInteractive Whether the output is interactive
+ * @param inputInteractive Whether the input is interactive
  */
 @JvmName("varargTest")
 fun CliktCommand.test(
@@ -74,15 +86,21 @@ fun CliktCommand.test(
     ansiLevel: AnsiLevel = AnsiLevel.NONE,
     width: Int = 79,
     height: Int = 24,
+    hyperlinks: Boolean = ansiLevel != AnsiLevel.NONE,
+    outputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
+    inputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
 ): CliktCommandTestResult {
-    return test(argv.asList(), stdin, envvars, includeSystemEnvvars, ansiLevel, width, height)
+    return test(
+        argv.asList(), stdin, envvars, includeSystemEnvvars, ansiLevel, width, height,
+        hyperlinks, outputInteractive, inputInteractive
+    )
 }
 
 /**
  * Test this command, returning a result that captures the output and result status code.
  *
- * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed with [print] or
- * [println] is not.
+ * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed
+ * with [print] or [println] is not.
  *
  * @param argv The command line to send to the command
  * @param stdin Content of stdin that will be read by prompt options. Multiple inputs should be separated by `\n`.
@@ -92,6 +110,9 @@ fun CliktCommand.test(
  * @param ansiLevel Defaults to no colored output; set to [AnsiLevel.TRUECOLOR] to include ANSI codes in the output.
  * @param width The width of the terminal, used to wrap text
  * @param height The height of the terminal
+ * @param hyperlinks Whether to enable hyperlink support in the terminal
+ * @param outputInteractive Whether the output is interactive
+ * @param inputInteractive Whether the input is interactive
  */
 fun CliktCommand.test(
     argv: Array<String>,
@@ -101,15 +122,21 @@ fun CliktCommand.test(
     ansiLevel: AnsiLevel = AnsiLevel.NONE,
     width: Int = 79,
     height: Int = 24,
+    hyperlinks: Boolean = ansiLevel != AnsiLevel.NONE,
+    outputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
+    inputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
 ): CliktCommandTestResult {
-    return test(argv.asList(), stdin, envvars, includeSystemEnvvars, ansiLevel, width, height)
+    return test(
+        argv.asList(), stdin, envvars, includeSystemEnvvars, ansiLevel, width, height,
+        hyperlinks, outputInteractive, inputInteractive
+    )
 }
 
 /**
  * Test this command, returning a result that captures the output and result status code.
  *
- * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed with [print] or
- * [println] is not.
+ * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed
+ * with [print] or [println] is not.
  *
  * @param argv The command line to send to the command
  * @param stdin Content of stdin that will be read by prompt options. Multiple inputs should be separated by `\n`.
@@ -119,6 +146,9 @@ fun CliktCommand.test(
  * @param ansiLevel Defaults to no colored output; set to [AnsiLevel.TRUECOLOR] to include ANSI codes in the output.
  * @param width The width of the terminal, used to wrap text
  * @param height The height of the terminal
+ * @param hyperlinks Whether to enable hyperlink support in the terminal
+ * @param outputInteractive Whether the output is interactive
+ * @param inputInteractive Whether the input is interactive
  */
 fun CliktCommand.test(
     argv: List<String>,
@@ -128,17 +158,21 @@ fun CliktCommand.test(
     ansiLevel: AnsiLevel = AnsiLevel.NONE,
     width: Int = 79,
     height: Int = 24,
+    hyperlinks: Boolean = ansiLevel != AnsiLevel.NONE,
+    outputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
+    inputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
 ): CliktCommandTestResult {
-    return test(argv, stdin, envvars, includeSystemEnvvars, ansiLevel, width, height) {
-        parse(it)
-    }
+    return test(
+        argv, stdin, envvars, includeSystemEnvvars, ansiLevel, width,
+        height, hyperlinks, outputInteractive, inputInteractive
+    ) { parse(it) }
 }
 
 /**
  * Test this command, returning a result that captures the output and result status code.
  *
- * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed with [print] or
- * [println] is not.
+ * Note that only output printed with [echo][CliktCommand.echo] will be captured. Anything printed
+ * with [print] or [println] is not.
  *
  * @param argv The command line to send to the command
  * @param stdin Content of stdin that will be read by prompt options. Multiple inputs should be separated by `\n`.
@@ -148,6 +182,9 @@ fun CliktCommand.test(
  * @param ansiLevel Defaults to no colored output; set to [AnsiLevel.TRUECOLOR] to include ANSI codes in the output.
  * @param width The width of the terminal, used to wrap text
  * @param height The height of the terminal
+ * @param hyperlinks Whether to enable hyperlink support in the terminal
+ * @param outputInteractive Whether the output is interactive
+ * @param inputInteractive Whether the input is interactive
  * @param parse The function to call to parse the command line and run the command
  */
 inline fun <T : BaseCliktCommand<T>> BaseCliktCommand<T>.test(
@@ -158,10 +195,15 @@ inline fun <T : BaseCliktCommand<T>> BaseCliktCommand<T>.test(
     ansiLevel: AnsiLevel = AnsiLevel.NONE,
     width: Int = 79,
     height: Int = 24,
+    hyperlinks: Boolean = ansiLevel != AnsiLevel.NONE,
+    outputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
+    inputInteractive: Boolean = ansiLevel != AnsiLevel.NONE,
     parse: (argv: List<String>) -> Unit,
 ): CliktCommandTestResult {
     var exitCode = 0
-    val recorder = TerminalRecorder(ansiLevel, width, height)
+    val recorder = TerminalRecorder(
+        ansiLevel, width, height, hyperlinks, outputInteractive, inputInteractive
+    )
     recorder.inputLines = stdin.split("\n").toMutableList()
     configureContext {
         val originalReader = envvarReader