+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added `CompletionGenerator` to manually generate completions for a command.
 - Added `Context.exitProcess` which you can use to prevent the process from exiting during tests.
 - Added core module that supports watchOS, tvOS, and wasmWasi targets and has no dependencies.
+- Added more options to `CliktCommand.test` to control the terminal interactivity. ([#517](https://github.com/ajalt/clikt/pull/517))
 
 ### Changed
 - In a subcommand with `argument().multiple()`, the behavior is now the same regardless of the value of `allowMultipleSubcommands`: if a token matches a subcommand name, it's now treated as a subcommand rather than a positional argument.